Change ESCourseSerializer to an ESModelSerializer
