+++ course_catalog/serializers.py
@@ -40,27 +40,9 @@ class CourseSerializer(serializers.ModelSerializer):
     Serializer for Course model
     """
 
-    instructors = serializers.SerializerMethodField()
-    topics = serializers.SerializerMethodField()
-    prices = serializers.SerializerMethodField()
-
-    def get_prices(self, course):
-        """
-        Get the prices for a course
-        """
-        return list(course.prices.values("price", "mode"))
-
-    def get_instructors(self, course):
-        """
-        Get a list of instructors for the course
-        """
-        return [" ".join([i.first_name, i.last_name]) for i in course.instructors.all()]
-
-    def get_topics(self, course):
-        """
-        Get the topic names for a course
-        """
-        return list(course.topics.values_list("name", flat=True))
+    instructors = CourseInstructorSerializer(read_only=True, many=True, allow_null=True)
+    topics = CourseTopicSerializer(read_only=True, many=True, allow_null=True)
+    prices = CoursePriceSerializer(read_only=True, many=True, allow_null=True)
 
     class Meta:
         model = Course
